Added models initialization to celery task
