+++ backend/.gitignore
@@ -1,2 +1,5 @@
 /.pytest_cache
 __pycache__/
+*.bak
+*.dat
+*.dir
