+++ backend/post_popularity_rate_celery/popularity_rate.py
@@ -4,9 +4,9 @@ from os import getenv
 from typing import List
 from asgiref.sync import async_to_sync
 
-from databases_manager.postgres_manager.models import Post, PostActions
+from databases_manager.postgres_manager.models import Post, PostActions, Base
 from databases_manager.postgres_manager.database_utils import get_session
-
+from databases_manager.postgres_manager.database import initialize_models, engine
 
 from sqlalchemy import select, text, and_
 from sqlalchemy.orm import selectinload
@@ -32,10 +32,11 @@ def sync_update_post_rate_task() -> None:
 
 async def update_post_rates() -> None:
     """Calculates new rate, shitfing out old actions."""
+    session = await get_session()
     try:
+        await initialize_models(engine=engine, Base=Base)
         now = datetime.utcnow()
 
-        session = await get_session()
         posts_raw = await session.execute(
             select(Post)
             .where((now - Post.last_updated) > EXPIRATION_INTERVAL)
@@ -51,7 +52,6 @@ async def update_post_rates() -> None:
             post.popularity_rate = calculate_new_rate(actions=actions)
 
         await session.commit()
-        await session.aclose()
     except Exception as e:
         await session.rollback()
         raise e
